13.3.5
Went into the models index.js and had to define the relationship between the Post and User table so sequelize can undertand it.

The code to define their relationship looked like this.
--

// create associations
User.hasMany(Post, {
  foreignKey: "user_id",
});

Post.belongsTo(User, {
  foreignKey: "user_id",
});
--.

This reads pretty well but to sum it up it says that the User table can have multiple post tables associated with it.  Then we in and let it know that Post can only belong to ONE user.  Both times we reference the fk which is user_id.