14.1.3
We are using the same module from last week and building on it.  The first thing we did was add a public folder with a nested stylesheet folder.   Inside of that we added a copy and pasted style sheet then

went into the server.js and added this
--
const path = require("path");

and

app.use(express.static(path.join(__dirname, "public")));
--.

These two things together tell the server.js file to use everything in the public folder on load.

Then we tested it out with
--
http://localhost:3001/stylesheets/style.css
--.