14.1.4
Added the HTML.

First we installed handlebars with this
--

npm install express-handlebars
--.

Then we required it in the server.js and put a few more lines in to get it set up.

server.js handlebars set up code
--

const exphbs = require('express-handlebars');
const hbs = exphbs.create({});

app.engine('handlebars', hbs.engine);
app.set('view engine', 'handlebars');
--.

Next we created a view folder and a layouts folder a put a main.handlebars inside it and gave it the HTML starter code.  Finally we went up one folder back into the views and added a homepage.handlebars file that just contained a single div letting you know you are on the homepage.  The file structure here is strict because everything here will go into layouts and apply itself to the {{body}} in the main.handlebars html.

Then we changed the name of our routes folder to controllers and updated the server.js to reflect this name change.  Finally we created a

home-routes.js in the controllers folder and put in this
--
const router = require('express').Router();

router.get('/', (req, res) => {
  res.render('homepage');
});

module.exports = router;
--.

Which tells it to render the homepage.handlebars on load.

Also we had to update the controllers/index.js to include this file.

With this
--

const homeRoutes = require('./home-routes.js');

router.use('/', homeRoutes);
--.