14.4.3
Created the first partial.  We made a folder in the root of views along side the layouts folder.  Inside it we created the pot-info.handlebars file that then we cut out the article section from the hompage and single-post handlebars since they are identical code.  We also removed post. in front of every variable because this file does now know what the post object is yet.

This is what we moved into post-info.handlerbars
--

<article class="post">
  <div class="title">
    <a href="{{post_url}}" target="_blank">{{title}}</a>
    <span>({{post_url}})</span>
  </div>
  <div class="meta">
    {{vote_count}}
    point(s) by
    {{user.username}}
    on
    {{created_at}}
    |
    <a href="/post/{{id}}">{{comments.length}} comment(s)</a>
  </div>
</article>
--.

Then in the hompage and single-post.handlerbars where we removed the articles

we added this
--

{{> post-info post }}
--.
As you can see here the first bit references the page and the second renames it post.  Also to console.log in handlebars simply type {{log this}} with this being whatever it is you are trying to log.